separate domain and repository
