+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/repository/RoleRepository.java
@@ -1,5 +1,6 @@
-package com.amigotrip.domain;
+package com.amigotrip.repository;
 
+import com.amigotrip.domain.Role;
 import org.springframework.data.repository.CrudRepository;
 
 /**
