+++ WEB-MAC-HACK-KI/src/main/java/com/amigotrip/web/ApiArticleController.java
@@ -1,7 +1,9 @@
 package com.amigotrip.web;
 
 import com.amigotrip.domain.*;
-import com.amigotrip.domain.*;
+import com.amigotrip.repository.LocalsArticleRepository;
+import com.amigotrip.repository.TravelerArticleRepository;
+import com.amigotrip.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
